knowledge_base/requirements.txt
openai>=1.0.0
instructor>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg[pool]>=3.1.0
networkx>=3.0.0
graspologic>=3.0.0
google-generativeai>=0.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0.0
httpx>=0.25.0
python-multipart>=0.0.6